Extract some IBFT network classes to the consensus/common package
for crosschain implementation
