+++ consensus/common/src/main/java/org/hyperledger/besu/consensus/common/network/ValidatorMulticaster.java
@@ -10,13 +10,16 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.hyperledger.besu.consensus.ibft.network;
+package org.hyperledger.besu.consensus.common.network;
 
-import org.hyperledger.besu.ethereum.p2p.rlpx.connections.PeerConnection;
+import org.hyperledger.besu.ethereum.core.Address;
+import org.hyperledger.besu.ethereum.p2p.rlpx.wire.MessageData;
 
-public interface PeerConnectionTracker {
+import java.util.Collection;
 
-  void add(final PeerConnection newConnection);
+public interface ValidatorMulticaster {
 
-  void remove(final PeerConnection removedConnection);
+  void send(final MessageData message);
+
+  void send(final MessageData message, final Collection<Address> blackList);
 }
