+++ consensus/ibft/src/main/java/org/hyperledger/besu/consensus/ibft/network/IbftMessageTransmitter.java
@@ -12,6 +12,7 @@
  */
 package org.hyperledger.besu.consensus.ibft.network;
 
+import org.hyperledger.besu.consensus.common.network.ValidatorMulticaster;
 import org.hyperledger.besu.consensus.ibft.ConsensusRoundIdentifier;
 import org.hyperledger.besu.consensus.ibft.messagedata.CommitMessageData;
 import org.hyperledger.besu.consensus.ibft.messagedata.PrepareMessageData;
