+++ consensus/ibft/src/test/java/org/hyperledger/besu/consensus/ibft/IbftGossipTest.java
@@ -15,11 +15,11 @@ package org.hyperledger.besu.consensus.ibft;
 import static com.google.common.collect.Lists.newArrayList;
 import static org.mockito.Mockito.verify;
 
+import org.hyperledger.besu.consensus.common.network.ValidatorMulticaster;
 import org.hyperledger.besu.consensus.ibft.messagedata.ProposalMessageData;
 import org.hyperledger.besu.consensus.ibft.messagedata.RoundChangeMessageData;
 import org.hyperledger.besu.consensus.ibft.messagewrappers.IbftMessage;
 import org.hyperledger.besu.consensus.ibft.network.MockPeerFactory;
-import org.hyperledger.besu.consensus.ibft.network.ValidatorMulticaster;
 import org.hyperledger.besu.crypto.SECP256K1.KeyPair;
 import org.hyperledger.besu.ethereum.core.Address;
 import org.hyperledger.besu.ethereum.core.AddressHelpers;
